Fix graceful termination of the proxy in cases when req is destroyed
This was a particular issue with the event stream keep alive connection
that webpack dev server creates. If req was destroyed, the proxyReq to
the webpack-dev-server stayed alive needlessly holding the server up.

This doesn't really affect production scenarios, but it's best practise
to clean resources up nicely even in dev.